+++ src/GraphCore/GraphState.hpp
@@ -64,12 +64,14 @@ public:
     void DrawCentralPoint(HDC hdc);
     void RenderGraph(HDC hdc);
     
+    Vec2d ExtractWindowSize();
     CursorType ExtractCursorType();
     bool ExtractFlagOfPlotResize();
 
     void AddData(std::vector<double>& load_data, std::wstring caption, COLORREF color, double step = 1, double offset = 0);
     size_t CreateTrace(std::wstring caption, COLORREF color, double step = 1.0, double offset = 0.0);
     bool StreamUpdate(std::vector<double>& load_data, size_t trace_index);
+    bool StreamAppend(std::vector<double>& load_data, size_t trace_index);
 };
 
 #endif